+++ packages/nextjs/src/proxy/index.ts
@@ -9,6 +9,7 @@ export {
   PersonalizeServiceConfig,
 } from '@sitecore-content-sdk/content/personalize';
 export { BotTrackingProxy, BotTrackingProxyConfig } from './bot-tracking-proxy';
+export { PreviewProxy, PreviewProxyConfig } from './preview-proxy';
 export {
   RedirectsService,
   RedirectsServiceConfig,