+++ CHANGELOG.md
@@ -14,17 +14,23 @@ Our versioning strategy is as follows:
 
 ### 🎉 New Features & Improvements
 
-* `[events]` `[nextjs]` Lightweight Tracking ([#414](https://github.com/Sitecore/content-sdk/pull/414))([#435](https://github.com/Sitecore/content-sdk/pull/435))([#438](https://github.com/Sitecore/content-sdk/pull/438))
+* `[events]` `[nextjs]` Lightweight Tracking ([#414](https://github.com/Sitecore/content-sdk/pull/414))([#435](https://github.com/Sitecore/content-sdk/pull/435))([#438](https://github.com/Sitecore/content-sdk/pull/438))([#443](https://github.com/Sitecore/content-sdk/pull/443))([#444](https://github.com/Sitecore/content-sdk/pull/444))
   - Introduced `BotTrackingProxy` Next.js proxy to capture bot tracking events.
 * `[nextjs]` Upgrade to Next.js 16.2 ([#429](https://github.com/Sitecore/content-sdk/pull/429))
 * `[create-content-sdk-app]` Use native flat ESLint config for App Router templates ([#431](https://github.com/Sitecore/content-sdk/pull/431))
 * `[nextjs]` Support for redirect items redirects in redirect proxy ([#433](https://github.com/Sitecore/content-sdk/pull/433)) ([#439](https://github.com/Sitecore/content-sdk/pull/439))
+* `[search]` Pass CDP session id in the queries to the search service ([#445](https://github.com/Sitecore/content-sdk/pull/445))
 
 ### 🐛 Bug Fixes
 
+* `[nextjs]` `[App Router]` searchParams empty on statically generated pages when Draft Mode is enabled (Vercel only) ([#448](https://github.com/Sitecore/content-sdk/pull/448))
+  - `searchParams` are not expected to be accessible in `draftMode` (this is a known Next.js issue). By design, preview data should be passed via request headers. To support this, we introduced the `client.getPreviewData` helper method. At the same time, preview data continues to be available via searchParams for backward compatibility. See more details in 'What's New' section of the release notes.
+* `[nextjs]` `[Internal Host]` Preview allows users to access pages without proper permissions ([#448](https://github.com/Sitecore/content-sdk/pull/448))([#455](https://github.com/Sitecore/content-sdk/pull/455))([#456](https://github.com/Sitecore/content-sdk/pull/456))
+  - App Router & Pages Router: Import and use `PreviewProxy` to gate preview requests. See more details in 'What's New' section of the release notes.
 * Preview mode shows unpublishable content ([#410](https://github.com/Sitecore/content-sdk/pull/410))([416](https://github.com/Sitecore/content-sdk/pull/416))
 * `[core] [content]` Fix GraphQL client factory ignoring custom `fetch` and related options ([#418](https://github.com/Sitecore/content-sdk/pull/418))
 * `[nextjs]` AppRouter - NextLink is throwing Locale error in dev mode ([#427](https://github.com/Sitecore/content-sdk/pull/427))
+* `[react]` Form component loses interactivity and state between rerenders ([#447](https://github.com/Sitecore/content-sdk/pull/447))
 
 ### ✨ Chores